models/segmentation_model/requirements.txt
numpy==1.21.5
flask==2.0.3
tensorflow==2.8.0
waitress==2.1.1
protobuf==3.20.*
pyyaml==6.0